requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
autopep8==2.3.1
backports-zoneinfo==0.2.1
black==24.8.0
cffi==1.17.1
click==8.1.8
django==4.2.23
django-admin-env-notice==1.0.1
django-admin-sortable2==2.2.8
django-ckeditor==6.7.3
django-cors-headers==4.4.0
django-csp==4.0
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-hijack==3.6.2
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-silk==5.2.0
django-stubs==5.1.3
django-stubs-ext==5.1.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.10
exceptiongroup==1.3.0
flake8==7.1.2
gprof2dot==2025.4.14
gunicorn==23.0.0
inflection==0.5.1
iniconfig==2.1.0
isort==5.13.2
mccabe==0.7.0
mypy==1.14.1
mypy-extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.3.6
pluggy==1.5.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
pyjwt==2.9.0
pytest==8.3.5
pytest-django==4.11.1
python-dotenv==1.0.1
pytz==2025.2
pyyaml==6.0.2
ruff==0.12.1
sqlparse==0.5.3
tomli==2.2.1
types-pyyaml==6.0.12.20241230
typing-extensions==4.13.2
uritemplate==4.1.1
whitenoise==6.7.0
